add modular history layers
