+++ src/ui/assignDialog.ts
@@ -1,5 +1,5 @@
 import { labelFromId } from '@/utils/names';
-import { findShiftsByStaff } from '@/state/history';
+import { findShiftsByStaff } from '@/history';
 import type { Staff } from '@/state';
 
 /**
