+++ src/ui/historyTab.ts
@@ -1,4 +1,4 @@
-import { renderHistory } from '@/history';
+import { renderHistory } from '@/history/ui';
 
 /** Legacy wrapper to render History tab. */
 export function renderHistoryTab(root: HTMLElement): void {
